mk: Remove use of /bin/ksh wrapper hack on Darwin.
There are only a few releases where this appears to work, and the commit logs
show this to be a complete pain to get working.  I've now confirmed it's also
broken on Big Sur where ksh simply segfaults on both x86_64 and arm64.

This may have been marginally useful in the past, but we now have cwrappers to
provide a much bigger performance improvement, and so this hack was only used
during bootstrap to build a small number of packages anyway.