audio/termusic: update to 0.6.0
-Update to tui-realm 1.0. Please note, as it's basically a rewrite, small bugs
are inevitable. Please report it in issues and I'll respond very fast.
-Hotkey to quit changed from Q to q, as now there will show a popup confirmation
so it's unlikely to quit accidentally.
-Can add a song to the beginning of playlist or the end. Switch by a.
Note: I add this to configuration file, and it'll reset your configuration file
to default values. Please backup if you need. Basically it's not necessary as
all options can be set from inside termusic.