open-vm-tools: updated to 11.3.5
open-vm-tools 11.3.5

For issues resolved in this release, see Resolved Issues section.

Added a configurable logging capability to the network script.

The network script has been updated to:

use vmware-toolbox-cmd to query any network logging configuration from the tools.conf file.
use vmtoolsd --cmd "log ..." to log a message to the vmx logfile when the logging handler is configured to "vmx" or when the logfile is full or is not writeable.
The hgfsmounter (mount.vmhgfs) command has been removed from open-vm-tools.

The hgfsmounter (mount.vmhgfs) command is no longer used in Linux open-vm-tools. It has been replaced by hgfs-fuse. Therefore, removing all references to the hgfsmounter in Linux builds.


Open-vm-tools 11.3.0

A small command line tool, vmwgfxctrl, has been added to open-vm-tools for Linux that can be used to control various aspects of the vmwgfx Linux kernel module. Currently it can both display and set the current topology of the vmwgfx kernel driver. It is useful when trying to configure custom resolutions on recent Linux distributions, including multi-monitor setups.
A command line tool, vmware-alias-import, has been added to open-vm-tools that can be used to import vgauth config data and apply it to the running vgauth service.