webkit-gtk: fix builds on NetBSD 9.x
The base GCC 7.5 on NetBSD 9.x is missing a C++17 feature expected by
this package now. The pkgsrc GCC 7.5 includes that feature, but the
build still fails later on with another unmet expectation. So GCC 8 is
now the minimum viable version with which to build this package.

This involves a bit of a kludge to deal with how GCC from pkgsrc ends
up linking. This is obviously not ideal, and is part of a broader
topic that needs revisiting. The present goal is to try and stabilize
the 2021Q4 branch.

Tested on 9.2_STABLE/amd64 with www/badwolf. Addresses PR pkg/56604.