textproc/git-delta: update to 0.11.3
0.9.0
=====
With this release, long lines are wrapped in side-by-side mode, thanks to a
far-reaching series of contributions by @th1000s (see #515). No special
customization is needed, but sophisticated control over the details is
available: see --wrap-max-lines, --wrap-left-symbol, --wrap-right-symbol,
--wrap-right-percent, --wrap-right-prefix-symbol, --inline-hint-style.

The long deleted line in the left panel overflows by a small amount, and the
wrapped content is right-aligned in the next line. In contrast, the long
replacement line in the right panel overflows by almost an entire line,
and so the wrapped content is left aligned in the next line. The arrow markers
and ellipsis explain when and how text has been wrapped.

What's Changed
-Fix the empty result issue in add -p by @norisio in #664
-Recognize GitHub SSH remote URLs that don't start with git@ for hyperlinks by
 @spaarmann in #668
-Terminate process gracefully on error in diff() by @dandavison in #685
-Ignore ctrl-c (SIGINT) to avoid leaving an orphaned pager process. by
 @dandavison in #686
-Remove unnecessary borrows by @benarmstead in #692
-Fix computed values to be computed after all set_options by @ulwlu in #690
-Make it possible to jump between files when navigate is active by @dandavison
 in #684
-Fix empty line highlighting by @th1000s in #642
-Fix deadlock in git diff mode by @dandavison in #695
-Support insteadOf replacements in git remote URLs by @dandavison in #694
-Do not suppress submodule diffs by @dandavison in #699
-More minimalistic display of submodule (short) diffs by @dandavison in #700
-Fix hunk header when line number is not requested and navigate is enabled by
 @dandavison in #710
-Display removed file name in file deletion case by @dandavison in #717
-Remove superfluous lowercasing of feature names by @Tak in #718
-Option to set the background extension mode to ANSI or spaces by @th1000s
 in #512
-Add side-by-side line wrapping mode by @th1000s in #515
-Width can be an offset from the terminal width by @th1000s in #727
-Linenumber refactoring by @th1000s in #740
-Use fatal() to exit with errorcode 2 by @th1000s in #739
-Re-enable ANSI fill by equalizing panel sizes by @th1000s in #742
-Update syntaxes and themes from bat v0.18.3 by @FnControlOption in #702
-Line numbers and side-by-side mode by @th1000s in #748

0.9.1
=====
This is a followup to release 0.9.0 fixing a bug in the --help text emitted by
the 0.9.0 binary (see 3aab5d1).

0.9.0 was a major release adding line-wrapping to side-by-side mode: 0.9.0
release notes.

0.9.2
=====
This release fixes a regression introduced in 0.9.0 affecting the use of
git add -p when submodules are present. See #755

What's Changed
-Do not reformat submodule diffs under git add -p by @dandavison in #756

0.10.0
======
New features in this release:

-delta displays output from git grep and rg --json ... | delta with code syntax
 highlighting and navigate functionality. See README#grep and #774.
-delta displays output from git blame with code syntax highlighting, hyperlinks
 to commits, and control over formatting. See README#git-blame and #761.
-delta syntax highlights file contents produced by git show $revision
 /path/to/file
-delta can apply syntax highlighting and arbitrary styling to the raw colors
 emitted by git --color-moved. This is done by specifying rules for remapping
 styles using map-styles. See below, and README#--color-moved-support and #72
-More powerful configuration options: styles can reference other styles, custom
 styles and custom colors can be defined in gitconfig and reused, and 140 colors
 can be specified using their standard CSS names. See below.
-New subcommand --show-colors outputs a demo of 140 background colors with
 syntax highlighting, to help choose background colors.
-New subcommand --parse-ansi takes color input and prints out the corresponding
 style strings (useful for map-styles)
-Utilities for interrogating the command line of the process that is sending
 output to delta, added by @th1000s.

Additionally, we can now use the 140 color names that are standard in CSS. Use
delta --show-colors to get a demo of the available colors, as background colors
to see how they look with syntax highlighting.

0.10.1
======
This release fixes a bug in color-moved handling when files have DOS line
endings. See #800

What's Changed
-Remove trailing \r from raw line, in presence of ANSI escape sequences by
 @dandavison in #800

0.10.2
======
This releases fixes a bug when using side-by-side with git diff.colorMoved.
See #802

What's Changed
-802 fix color moved side by side panic by @dandavison in #803

0.10.3
======
Delta was previously handling --word-diff and --color-words output in a
nonsensical way. With this release, delta does not attempt to handle these
output types, allowing git's raw output through unchanged. Appropriate handling
may be added later. See #152
When a line is unchanged but paired with a line that has additions/deletions,
delta now assigns the unchanged line {plus,minus}-non-emph-style; previously it
assigned {plus,minus}-style. This change does not alter visual output by default
because by default those two style styles are set to the same value. See #785
Fixes a bug affecting line numbers in side-by-side mode, see #819

0.11.0
======
This release adds handling of merge conflicts and the "combined" diff format
which git uses for merge commits; see the Merge conflicts README section and
#189. Please consider setting

[merge]
    conflictstyle = diff3

With that setting, when a merge conflict is encountered, delta will display
diffs between the ancestral commit and each of the two merge parents.

0.11.1
======
Fixes a crash that can occur, see for example #826. Thanks @mrjoel.

What's Changed
-Be permissive about unexpected state machine state by @dandavison in #827

0.11.2
======
This release fixes a performance problem that was affecting some users. See #824

0.11.3
======
Fix handling of --word-diff and --color-words (#829). Note that line-numbers and
side-by-side will be ignored if delta detects that the input was generated by
git --word-diff / --color-words.
DELTA_FEATURES can now be used to add to features specified in ~/.gitconfig.
Use a + prefix to do this: an example is DELTA_FEATURES=+side-by-side which will
activate side-by-side, in addition to your usual features.
Store less history file in a better location on Windows. Thanks @rashil2000 for
this work!

Fix performance issue affecting some linux users (#839)

What's Changed
-Bump sysinfo from 0.19.2 to 0.20.5 by @dependabot in #811
-#833 refactoring by @dandavison in #834
-Store history file in data_local_dir for Windows by @rashil2000 in #847
-Query fewer processes when searching for the parent by @th1000s in #841
-Do not query CPU data when querying process data by @dandavison in #845
-Fix --word-diff / --color-words handling by @dandavison in #833
-Add features from DELTA_FEATURES env var instead of replacing with them by
 @dandavison in #849
-Fix raw line styles by @dandavison in #850
-Fix side-by-side syntax ANSI test failure by @dandavison in #852
-DeltaTest inspect() additions by @dandavison in #853
-Line state refactor by @dandavison in #851