gopls: update to 0.7.5.
Configuration changes

templateExtensions now defaults to []

Following feedback on the template features introduced with v0.7.4,
we have made improvements to the way the active set of template
files is configured. The templateExtensions setting is now empty
by default, and gopls recognizes any open files with the "gotmpl"
language identifier as a template file. As a result, unopened files
will not be loaded as Go templates unless the templateExtensions
is configured to be non-empty, but gopls will still provide template
language features to any open files configured to be Go template
files by the LSP client.

As always, your feedback helps make gopls better, and is greatly
appreciated!

Features

New postfix completions

    The split! completion on values of string type inserts a
    strings.Split snippet.  split

    The join! completion on values of []string type inserts a
    strings.Join snippet.  join

New suggested fix for unused function parameters

The diagnostic for unused function parameters now includes a
suggested fix to rename the parameter to _. This diagnostic is
disabled by default, but can be enabled via the unusedparams setting
in the analyses section.